requirements.txt
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
pillow==11.1.0
PyYAML==6.0.2
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
typing_extensions==4.13.1
optuna==4.3.0
